+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added libcasm.configuration.Prim.lattice_point_group
 - Added libcasm.configuration.io.spglib to interface with spglib
 - Added libcasm.io.symgroup_to_dict_with_group_classification to add group classification from spglib to SymGroup.to_dict data.
+- Added libcasm.clusterography.make_local_cluster_group
 
 
 ## [v2.0a3] - 2024-03-15